Change coin key

+++ WalletWasabi.Fluent/ViewModels/CoinSelection/CoinBasedSelectionViewModel.cs
@@ -19,7 +19,7 @@ public partial class CoinBasedSelectionViewModel : ViewModelBase, IDisposable
 	[AutoNotify(SetterModifier = AccessModifier.Private)]
 	private HierarchicalTreeDataGridSource<TreeNode> _source;
 
-	public CoinBasedSelectionViewModel(IObservable<IChangeSet<WalletCoinViewModel, uint256>> coinChanges)
+	public CoinBasedSelectionViewModel(IObservable<IChangeSet<WalletCoinViewModel, int>> coinChanges)
 	{
 		coinChanges
 			.Transform(model => new TreeNode(model))
